Added Straight Line Distance-heuristic

+++ src/pathfinding/IHeuristic.java
@@ -2,6 +2,6 @@ package pathfinding;
 
 public interface IHeuristic {
 
-	public float getHeuristicDistance(Node current, Node goal);
+	public double getHeuristicDistance(Node current, Node goal);
 	
 }
